Documentation

Index

A list of all documentation sorted by module.

QuantumInformation.QuantumInformation

QuantumInformation.ChannelBasisIterator

QuantumInformation.DynamicalMatrix

QuantumInformation.IdentityChannel

QuantumInformation.KrausOperators

QuantumInformation.Stinespring

QuantumInformation.SuperOperator

QuantumInformation.SuperOperator

QuantumInformation.UnitaryChannel

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.bloch_vector

QuantumInformation.bra

QuantumInformation.bures_angle

QuantumInformation.bures_distance

QuantumInformation.channelbasis

QuantumInformation.combine

QuantumInformation.concurrence

QuantumInformation.diamond_distance

QuantumInformation.fidelity

QuantumInformation.fidelity_sqrt

QuantumInformation.gate_fidelity

QuantumInformation.grover

QuantumInformation.hadamard

QuantumInformation.hermitianbasis

QuantumInformation.hs_distance

QuantumInformation.iscp

QuantumInformation.iscptni

QuantumInformation.iscptp

QuantumInformation.iseffect

QuantumInformation.isidentity

QuantumInformation.ispositive

QuantumInformation.ispovm

QuantumInformation.istni

QuantumInformation.istp

QuantumInformation.js_divergence

QuantumInformation.ket

QuantumInformation.ketbra

QuantumInformation.ketbra

QuantumInformation.kl_divergence

QuantumInformation.log_negativity

QuantumInformation.max_entangled

QuantumInformation.max_mixed

QuantumInformation.mixedradix2number

QuantumInformation.negativity

QuantumInformation.norm_diamond

QuantumInformation.norm_hs

QuantumInformation.norm_trace

QuantumInformation.number2mixedradix

QuantumInformation.permutesystems

QuantumInformation.ppt

QuantumInformation.proj

QuantumInformation.ptrace

QuantumInformation.ptrace

QuantumInformation.ptrace

QuantumInformation.ptranspose

QuantumInformation.ptranspose

QuantumInformation.purity

QuantumInformation.qft

QuantumInformation.relative_entropy

QuantumInformation.renormalize!

QuantumInformation.renormalize!

QuantumInformation.renyi_entropy

QuantumInformation.represent

QuantumInformation.res

QuantumInformation.reshuffle

QuantumInformation.shannon_entropy

QuantumInformation.shannon_entropy

QuantumInformation.superfidelity

QuantumInformation.trace_distance

QuantumInformation.unres

QuantumInformation.vonneumann_entropy

QuantumInformation.werner_state

QuantumInformation.QuantumInformation — Module

Main module for QuantumInformation.jl – a Julia package for numerical computation in quantum information theory.

QuantumInformation.ChannelBasisIterator — Type

QuantumInformation.DynamicalMatrix — Type

T: quantum channel map.

Representation of quantum channel by Dynamical matrix operators.

QuantumInformation.IdentityChannel — Type

T: quantum channel map.

Representation of identity channel.

QuantumInformation.KrausOperators — Type

T: quantum channel map.

Representation of quantum channel by Kraus operators.

QuantumInformation.Stinespring — Type

T: quantum channel map.

Stinespring representation of quantum channel.

QuantumInformation.SuperOperator — Type

T: quantum channel map.

Representation of quantum channel by super-operator.

QuantumInformation.SuperOperator — Method

channel: quantum channel map.

idim: square root of the super-operator matrix input dimension.

odim: square root of the super-operator matrix output dimension.

Transforms quntum channel into super-operator matrix.

QuantumInformation.UnitaryChannel — Type

T: quantum channel map.

Representation of unitary channel.

Base.convert — Method

?: type.

Φ: list of Kraus operators.

Transforms list of Kraus operators into dynamical matrix.

Base.convert — Method

?: type.

Φ: super-operator matrix.

Transforms super-operator matrix into dynamical matrix.

Base.convert — Method

?: type.

Φ: dynamical matrix.

Transforms dynamical matrix into list of Kraus operators.

Base.convert — Method

?: type.

Φ: super-operator matrix.

Transforms super-operator matrix into list of Kraus operators.

Base.convert — Method

?: type.

Φ: dynamical matrix.

Transforms dynamical matrix into Stinespring representation of quantum channel.

Base.convert — Method

?: type.

Φ: list of Kraus operators.

Transforms list of Kraus operators into Stinespring representation of quantum channel.

Base.convert — Method

?: type.

Φ: super-operator matrix.

Transforms super-operator matrix into Stinespring representation of quantum channel.

Base.convert — Method

?: type.

Φ: dynamical matrix.

Transforms dynamical matrix into super-operator matrix.

Base.convert — Method

?: type.

Φ: list of Kraus operators.

Transforms list of Kraus operators into super-operator matrix.

QuantumInformation.applychannel — Method

Φ: quantum channel.

ψ: quantum state vector.

Return application of channel Φ on state vector ψ.

QuantumInformation.applychannel — Method

Φ: Identity channel.

ρ: quantum state.

Return application of Identity channel Φ on ρ.

QuantumInformation.applychannel — Method

Φ: Identity channel.

ψ: quantum state vector.

Return application of Identity channel Φ on state vector ψ.

QuantumInformation.applychannel — Method

Φ: list of vectors.

ρ: input matrix.

Return application of channel Φonρ`. Kraus representation of quantum channel $\Phi$ is a set $\{K_i\}_{i\in I}$ of bounded operators on $\mathcal{H}$ such that $\sum_{i\in I} K_i^\dagger K_i = \mathcal{1}$. Then $\Phi(\rho)=\sum_{i\in I} K_i \rho K_i^\dagger$.

QuantumInformation.applychannel — Method

Φ: Stinespring representation of quantum channel.

ρ: quantum state.

dims: dimensions of registers of ρ.

Application of Stinespring representation of quantum channel into state ρ.

QuantumInformation.applychannel — Method

Φ: super-operator matrix.

ρ: quantum state.

Application of super-operator matrix into state ρ.

QuantumInformation.applychannel — Method

Φ: Unitary channel.

ρ: quantum state.

Return application of Unitary channel Φ on ρ.

QuantumInformation.applychannel — Method

Φ: Unitary channel.

ψ: quantum state vector.

Return application of Unitary channel Φ on state vector ψ.

QuantumInformation.applychannel — Method

Φ: dynamical matrix.

ρ: quantum state.

Application of dynamical matrix into state ρ.

QuantumInformation.bloch_vector — Method

ρ: input qubit density matrix.

Return the Bloch vector corresponding to the inpu quit state.

QuantumInformation.bra — Method

val: non-zero entry - label.

dim: length of the vector

Return Hermitian conjugate $\langle val| = |val\rangle^\dagger$ of the ket with the same label.

QuantumInformation.bures_angle — Method

ρ: quantum state.

σ: quantum state.

Return Bures angle between quantum states ρ and σ.

QuantumInformation.bures_distance — Method

ρ: quantum state.

σ: quantum state.

Return Bures distance between quantum states ρ and σ.

QuantumInformation.channelbasis — Function

T: Type of the matrices in the basis.

idim: Input dimension.

odim: Output dimension.

Returns a basis for quantum channels.

QuantumInformation.combine — Method

basis: A basis represented by a sub-type of AbstractMatrixBasis.

v: Vector of coefficients.

Returns a matrix constructed from the basis elements weighted by the coefficients in v.

QuantumInformation.concurrence — Method

ρ: quantum state.

Calculates the concurrence of a two-qubit system ρ.

QuantumInformation.diamond_distance — Method

Φ1: DynamicalMatrix

Φ2: DynamicalMatrix

Return diamond distance between dynamical matrices Φ1 and Φ2.

QuantumInformation.fidelity — Method

ρ: matrix.

σ: matrix.

Return fidelity between matrices ρ and σ.

QuantumInformation.fidelity_sqrt — Method

ρ: matrix.

σ: matrix.

Return square root of fidelity between matrices ρ and σ.

QuantumInformation.gate_fidelity — Method

U: quantum gate.

V: quantum gate.

Return fidelity between gates U and V.

QuantumInformation.grover — Method

d: dimension of operator.

Prepares Grover operator of dimension d.

QuantumInformation.hadamard — Method

d: dimension of operator.

Prepares Hadamard operator of dimension d.

QuantumInformation.hermitianbasis — Method

dim: dimensions of the matrix. Returns elementary hermitian matrices of dimension dim x dim.

QuantumInformation.hs_distance — Method

A: matrix.

B: matrix.

Return Hilbert–Schmidt distance between matrices A and B.

QuantumInformation.iscp — Function

Φ: A subtype of AbstractQuantumOperation.

atol: tolerance of approximation.

Checks if an object is completely positive.

QuantumInformation.iscptni — Function

Φ: A subtype of AbstractQuantumOperation.

atol: tolerance of approximation.

Checks if an object is completely positive and trace non-increasing.

QuantumInformation.iscptp — Function

Φ: A subtype of AbstractQuantumOperation.

atol: tolerance of approximation.

Checks if an object is completely positive and trace preserving.

QuantumInformation.iseffect — Method

Φ: Post-selection Measurement.

Checks if a matrix represents a valid quantum effect (0 <= E <= I).

QuantumInformation.isidentity — Function

ρ: Input matrix.

atol: Absolute tolerance.

Checks if the matrix ρ is approximately the identity matrix.

QuantumInformation.ispositive — Function

ρ: Input matrix.

atol: Absolute tolerance.

Checks if the matrix ρ is positive semi-definite.

QuantumInformation.ispovm — Method

Φ: POVM Measurement.

Checks if a set of matrices forms a valid POVM (Positive Operator-Valued Measure).

QuantumInformation.istni — Function

Φ: A subtype of AbstractQuantumOperation.

atol: tolerance of approximation.

Checks if an object is trace non-increasing.

QuantumInformation.istp — Function

Φ: A subtype of AbstractQuantumOperation.

atol: tolerance of approximation.

Checks if an object is trace preserving.

QuantumInformation.js_divergence — Method

ρ: quantum state.

σ: quantum state.

Return Jensen–Shannon divergence of quantum state ρ with respect to σ.

QuantumInformation.ket — Method

val: non-zero entry - label.

dim: length of the vector.

Return complex column vector $|val\rangle$ of unit norm describing quantum state.

QuantumInformation.ketbra — Method

valk: non-zero entry - label.

valb: non-zero entry - label.

idim: length of the ket vector

odim: length of the bra vector

Return outer product $|valk\rangle\langle vakb|$ of states $|valk\rangle$ and $|valb\rangle$.

QuantumInformation.ketbra — Method

valk: non-zero entry - label.

valb: non-zero entry - label.

dim: length of the ket and bra vectors

Return outer product $|valk\rangle\langle vakb|$ of states $|valk\rangle$ and $|valb\rangle$.

QuantumInformation.kl_divergence — Method

ρ: quantum state.

σ: quantum state.

Return Kullback–Leibler divergence of quantum state ρ with respect to σ.

QuantumInformation.log_negativity — Method

ρ: quantum state.

dims: dimensions of subsystems.

sys: transposed subsystem.

Return log negativity of quantum state ρ.

QuantumInformation.max_entangled — Method

d: length of the vector.

Return maximally entangled state $\frac{1}{\sqrt{d}}\sum_{i=0}^{\sqrt{d}-1}|ii\rangle$ of length $\sqrt{d}$.

QuantumInformation.max_mixed — Method

d: length of the vector.

Return maximally mixed state $\frac{1}{d}\sum_{i=0}^{d-1}|i\rangle\langle i |$ of length $d$.

QuantumInformation.mixedradix2number — Method

digits: Vector of coefficients in mixed radix representation.

radices: Vector of mixed radices.

Returns the integer number corresponding to the mixed radix representation.

QuantumInformation.negativity — Method

ρ: quantum state.

dims: dimensions of subsystems.

sys: transposed subsystem.

Return negativity of quantum state ρ.

QuantumInformation.norm_diamond — Function

Φ: DynamicalMatrix

Return diamond norm of dynamical matrix Φ.

QuantumInformation.norm_hs — Method

A: matrix.

Return Hilbert–Schmidt norm of matrix A.

QuantumInformation.norm_trace — Method

A: matrix.

Return trace norm of matrix A.

QuantumInformation.number2mixedradix — Method

n: Number to be converted (integer).

radices: Vector of mixed radices.

Returns the representation of n in the mixed radix system defined by radices.

QuantumInformation.permutesystems — Method

ρ: input state.

dims: dimensions of registers of ρ.

systems: permuted registers.

Returns state ρ with permuted registers denoted by systems.

QuantumInformation.ppt — Method

ρ: quantum state.

dims: dimensions of subsystems.

sys: transposed subsystem.

Return minimum eigenvalue of positive partial transposition of quantum state ρ.

QuantumInformation.proj — Method

ket: input column vector.

Return outer product $|ket\rangle\langle ket|$ of ket.

QuantumInformation.ptrace — Method

ρ: quantum state.

idims: dimensins of subsystems.

sys: traced subsystem.

QuantumInformation.ptrace — Method

ρ: quantum state.

idims: dimensins of subsystems.

isystems: traced subsystems.

Return partial trace of matrix ρ over the subsystems determined by isystems.

QuantumInformation.ptrace — Method

ψ: quantum state pure state (ket).

idims: dimensins of subsystems - only bipartite states accepted.

sys: traced subsystem.

QuantumInformation.ptranspose — Method

ρ: quantum state.

idims: dimensins of subsystems.

sys: transposed subsystem.

QuantumInformation.ptranspose — Method

ρ: quantum state.

idims: dimensins of subsystems.

isystems: transposed subsystems.

Return partial transposition of matrix ρ over the subsystems determined by isystems.

QuantumInformation.purity — Method

ρ: matrix.

Return the purity of ρ ∈ [1/d, 1]

QuantumInformation.qft — Method

d: dimension of operator.

Prepares gate realized a quantum Fourier transform of dimension d.

QuantumInformation.relative_entropy — Method

ρ: quantum state.

σ: quantum state.

Return quantum relative entropy of quantum state ρ with respect to σ.

QuantumInformation.renormalize! — Method

ρ: Input matrix.

Renormalizes the matrix ρ in-place so that its trace is 1.

QuantumInformation.renormalize! — Method

ψ: Input vector.

Renormalizes the vector ψ in-place so that its norm is 1.

QuantumInformation.renyi_entropy — Method

ρ: quantum state.

α: order of Renyi entropy.

Return Renyi entropy of quantum state ρ.

QuantumInformation.represent — Method

basis: A basis represented by a sub-type of AbstractMatrixBasis.

m: Matrix to be represented in the basis.

Returns a vector of coefficients of the matrix m in the basis basis.

QuantumInformation.res — Method

ρ: input matrix.

Returns vec(ρ.T). Reshaping maps matrix ρ into a vector row by row.

QuantumInformation.reshuffle — Method

ρ: reshuffled matrix. Performs reshuffling of indices of a matrix. Given multiindexed matrix $M_{(m,μ),(n,ν)}$ it returns matrix $M_{(m,n),(μ,ν)}$.

QuantumInformation.shannon_entropy — Method

x: real number.

Return binary Shannon entorpy given by $-x \log(x) - (1 - x) \log(1 - x)$.

QuantumInformation.shannon_entropy — Method

p: vector.

Return Shannon entorpy of vector p.

QuantumInformation.superfidelity — Method

ρ: quantum state.

σ: quantum state.

Return superfidelity between quantum states ρ and σ.

QuantumInformation.trace_distance — Method

A: matrix.

B: matrix.

Return trace distance between matrices A and B.

QuantumInformation.unres — Method

ϕ: input matrix.

Return de-reshaping of the vector into a matrix.

QuantumInformation.vonneumann_entropy — Method

ρ: quantum state.

Return Von Neumann entropy of quantum state ρ.

QuantumInformation.werner_state — Method

d: length of the vector.

α: real number from [0, 1].

Returns Werner state given by $\frac{\alpha}{d}\left(\sum_{i=0}^{\sqrt{d}-1}|ii\rangle\right) \left(\sum_{i=0}^{\sqrt{d}-1}\langle ii|\right)+ \frac{1-\alpha}{d}\sum_{i=0}^{d-1}|i\rangle\langle i|$.