Library

Documentation

Index

A list of all documentation sorted by module.

QuantumInformation.QuantumInformation

QuantumInformation.DynamicalMatrix

QuantumInformation.HaarKet

QuantumInformation.HilbertSchmidtStates

QuantumInformation.IdentityChannel

QuantumInformation.KrausOperators

QuantumInformation.Stinespring

QuantumInformation.SuperOperator

QuantumInformation.SuperOperator

QuantumInformation.UnitaryChannel

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

Base.convert

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.applychannel

QuantumInformation.bra

QuantumInformation.bures_angle

QuantumInformation.bures_distance

QuantumInformation.concurrence

QuantumInformation.diamond_distance

QuantumInformation.fidelity

QuantumInformation.fidelity_sqrt

QuantumInformation.gate_fidelity

QuantumInformation.grover

QuantumInformation.hadamard

QuantumInformation.hermitianbasis

QuantumInformation.hs_distance

QuantumInformation.iscptp

QuantumInformation.js_divergence

QuantumInformation.ket

QuantumInformation.ketbra

QuantumInformation.ketbra

QuantumInformation.kl_divergence

QuantumInformation.log_negativity

QuantumInformation.max_entangled

QuantumInformation.max_mixed

QuantumInformation.negativity

QuantumInformation.norm_diamond

QuantumInformation.norm_hs

QuantumInformation.norm_trace

QuantumInformation.permutesystems

QuantumInformation.ppt

QuantumInformation.proj

QuantumInformation.ptrace

QuantumInformation.ptrace

QuantumInformation.ptrace

QuantumInformation.ptranspose

QuantumInformation.ptranspose

QuantumInformation.purity

QuantumInformation.qft

QuantumInformation.relative_entropy

QuantumInformation.res

QuantumInformation.reshuffle

QuantumInformation.shannon_entropy

QuantumInformation.shannon_entropy

QuantumInformation.superfidelity

QuantumInformation.trace_distance

QuantumInformation.unres

QuantumInformation.vonneumann_entropy

QuantumInformation.werner_state

QuantumInformation.QuantumInformation — Module.

Main module for QuantumInformation.jl – a Julia package for numerical computation in quantum information theory.

source

QuantumInformation.DynamicalMatrix — Type.

T: quantum channel map.

Representation of quantum channel by Dynamical matrix operators.

source

QuantumInformation.HaarKet — Type.

ϕ: vector.

Generates random ket based on ϕ.

source

QuantumInformation.HilbertSchmidtStates — Type.

d: length.

Generates random ket of length d.

source

QuantumInformation.IdentityChannel — Type.

T: quantum channel map.

Representation of identity channel.

source

QuantumInformation.KrausOperators — Type.

T: quantum channel map.

Representation of quantum channel by Kraus operators.

source

QuantumInformation.Stinespring — Type.

T: quantum channel map.

Stinespring representation of quantum channel.

source

QuantumInformation.SuperOperator — Type.

T: quantum channel map.

Representation of quantum channel by super-operator.

source

QuantumInformation.SuperOperator — Method.

SuperOperator(m)

channel: quantum channel map.

dim: square root of the super-operator matrix dimension.

Transforms quntum channel into super-operator matrix.

source

QuantumInformation.UnitaryChannel — Type.

T: quantum channel map.

Representation of unitary channel.

source

QuantumInformation.applychannel — Method.

applychannel(Φ, ρ)

Φ: list of vectors.

ρ: input matrix.

Return application of channel Φonρ`. Kraus representation of quantum channel $\Phi$ is a set $\{K_i\}_{i\in I}$ of bounded operators on $\mathcal{H}$ such that $\sum_{i\in I} K_i^\dagger K_i = \mathcal{1}$. Then $\Phi(\rho)=\sum_{i\in I} K_i \rho K_i^\dagger$.

source

QuantumInformation.applychannel — Method.

applychannel(Φ, ρ)

Φ: Stinespring representation of quantum channel.

ρ: quantum state.

dims: dimensions of registers of ρ.

Application of Stinespring representation of quantum channel into state ρ.

source

QuantumInformation.applychannel — Method.

applychannel(Φ, ρ)

Φ: super-operator matrix.

ρ: quantum state.

Application of super-operator matrix into state ρ.

source

QuantumInformation.applychannel — Method.

applychannel(Φ, ρ)

Φ: dynamical matrix.

ρ: quantum state.

Application of dynamical matrix into state ρ.

source

QuantumInformation.bra — Method.

bra(val, dim)

val: non-zero entry - label.

dim: length of the vector

Return Hermitian conjugate $\langle val| = |val\rangle^\dagger$ of the ket with the same label.

source

QuantumInformation.bures_angle — Method.

bures_angle(ρ, σ)

ρ: quantum state.

σ: quantum state.

Return Bures angle between quantum states ρ and σ.

source

QuantumInformation.bures_distance — Method.

bures_distance(ρ, σ)

ρ: quantum state.

σ: quantum state.

Return Bures distance between quantum states ρ and σ.

source

QuantumInformation.concurrence — Method.

concurrence(ρ)

ρ: quantum state.

Calculates the concurrence of a two-qubit system ρ.

source

QuantumInformation.diamond_distance — Method.

diamond_distance(Φ1, Φ2, args)

Φ1: DynamicalMatrix

Φ2: DynamicalMatrix

Return diamond distance between dynamical matrices Φ1 and Φ2.

source

QuantumInformation.fidelity — Method.

fidelity(ρ, σ)

ρ: matrix.

σ: matrix.

Return fidelity between matrices ρ and σ.

source

QuantumInformation.fidelity_sqrt — Method.

fidelity_sqrt(ρ, σ)

ρ: matrix.

σ: matrix.

Return square root of fidelity between matrices ρ and σ.

source

QuantumInformation.gate_fidelity — Method.

gate_fidelity(U, V)

U: quantum gate.

V: quantum gate.

Return fidelity between gates U and V.

source

QuantumInformation.grover — Method.

grover(dim)

d: dimension of operator.

Prepares Grover operator of dimension d.

source

QuantumInformation.hadamard — Method.

hadamard(dim)

d: dimension of operator.

Prepares Hadamard operator of dimension d.

source

QuantumInformation.hermitianbasis — Method.

hermitianbasis(T, dim)

dim: dimensions of the matrix.

Returns elementary hermitian matrices of dimension dim x dim.

source

QuantumInformation.hs_distance — Method.

hs_distance(A, B)

A: matrix.

B: matrix.

Return Hilbert–Schmidt distance between matrices A and B.

source

QuantumInformation.iscptp — Method.

iscptp(Φ; atol)

Φ: list of Kraus operators.

atol: tolerance of approximation.

Checks if set of Kraus operators fulfill completness relation.

source

QuantumInformation.js_divergence — Method.

js_divergence(ρ, σ)

ρ: quantum state.

σ: quantum state.

Return Jensen–Shannon divergence of quantum state ρ with respect to σ.

source

QuantumInformation.ket — Method.

ket(val, dim)

val: non-zero entry - label.

dim: length of the vector.

Return complex column vector $|val\rangle$ of unit norm describing quantum state.

source

QuantumInformation.ketbra — Method.

valk: non-zero entry - label.

valb: non-zero entry - label.

dimin: length of the ket vector

dimout: length of the bra vector

Return outer product $|valk\rangle\langle vakb|$ of states $|valk\rangle$ and $|valb\rangle$.

source

QuantumInformation.ketbra — Method.

ketbra(valk, valb, dim)

valk: non-zero entry - label.

valb: non-zero entry - label.

dim: length of the ket and bra vectors

Return outer product $|valk\rangle\langle vakb|$ of states $|valk\rangle$ and $|valb\rangle$.

source

QuantumInformation.kl_divergence — Method.

kl_divergence(ρ, σ)

ρ: quantum state.

σ: quantum state.

Return Kullback–Leibler divergence of quantum state ρ with respect to σ.

source

QuantumInformation.log_negativity — Method.

log_negativity(ρ, dims, sys)

ρ: quantum state.

dims: dimensions of subsystems.

sys: transposed subsystem.

Return log negativity of quantum state ρ.

source

QuantumInformation.max_entangled — Method.

max_entangled(d)

d: length of the vector.

Return maximally entangled state $\frac{1}{\sqrt{d}}\sum_{i=0}^{\sqrt{d}-1}|ii\rangle$ of length $\sqrt{d}$.

source

QuantumInformation.max_mixed — Method.

max_mixed(d)

d: length of the vector.

Return maximally mixed state $\frac{1}{d}\sum_{i=0}^{d-1}|i\rangle\langle i |$ of length $d$.

source

QuantumInformation.negativity — Method.

negativity(ρ, dims, sys)

ρ: quantum state.

dims: dimensions of subsystems.

sys: transposed subsystem.

Return negativity of quantum state ρ.

source

QuantumInformation.norm_diamond — Method.

norm_diamond(Φ)
norm_diamond(Φ, method)
norm_diamond(Φ, method, eps)

Φ: DynamicalMatrix

Return diamond norm of dynamical matrix Φ.

source

QuantumInformation.norm_hs — Method.

norm_hs(A)

A: matrix.

Return Hilbert–Schmidt norm of matrix A.

source

QuantumInformation.norm_trace — Method.

norm_trace(A)

A: matrix.

Return trace norm of matrix A.

source

QuantumInformation.permutesystems — Method.

permutesystems(ρ, dims, systems)

ρ: input state.

dims: dimensions of registers of ρ.

systems: permuted registers.

Returns state ρ with permuted registers denoted by systems.

source

QuantumInformation.ppt — Method.

ppt(ρ, dims, sys)

ρ: quantum state.

dims: dimensions of subsystems.

sys: transposed subsystem.

Return minimum eigenvalue of positive partial transposition of quantum state ρ.

source

QuantumInformation.proj — Method.

proj(ψ)

ket: input column vector.

Return outer product $|ket\rangle\langle ket|$ of ket.

source

QuantumInformation.ptrace — Method.

ptrace(ψ, idims, sys)

ψ: quantum state pure state (ket).

idims: dimensins of subsystems - only bipartite states accepted.

sys: traced subsystem.

source

QuantumInformation.ptrace — Method.

ptrace(ρ, idims, isystems)

ρ: quantum state.

idims: dimensins of subsystems.

isystems: traced subsystems.

Return partial trace of matrix ρ over the subsystems determined by isystems.

source

QuantumInformation.ptrace — Method.

ptrace(ρ, idims, sys)

ρ: quantum state.

idims: dimensins of subsystems.

sys: traced subsystem.

source

QuantumInformation.ptranspose — Method.

ptranspose(ρ, idims, isystems)

ρ: quantum state.

idims: dimensins of subsystems.

isystems: transposed subsystems.

Return partial transposition of matrix ρ over the subsystems determined by isystems.

source

QuantumInformation.ptranspose — Method.

ptranspose(ρ, idims, sys)

ρ: quantum state.

idims: dimensins of subsystems.

sys: transposed subsystem.

source

QuantumInformation.purity — Method.

purity(ρ)

ρ: matrix.

Return the purity of ρ ∈ [1/d, 1]

source

QuantumInformation.qft — Method.

qft(d)

d: dimension of operator.

Prepares gate realized a quantum Fourier transform of dimension d.

source

QuantumInformation.relative_entropy — Method.

relative_entropy(ρ, σ)

ρ: quantum state.

σ: quantum state.

Return quantum relative entropy of quantum state ρ with respect to σ.

source

QuantumInformation.res — Method.

res(ρ)

ρ: input matrix.

Returns vec(ρ.T). Reshaping maps matrix ρ into a vector row by row.

source

QuantumInformation.reshuffle — Method.

reshuffle(ρ)

ρ: reshuffled matrix.

Performs reshuffling of indices of a matrix. Given multiindexed matrix $M_{(m,μ),(n,ν)}$ it returns matrix $M_{(m,n),(μ,ν)}$.

source

QuantumInformation.shannon_entropy — Method.

shannon_entropy(p)

p: vector.

Return Shannon entorpy of vector p.

source

QuantumInformation.shannon_entropy — Method.

shannon_entropy(x)

x: real number.

Return binary Shannon entorpy given by $-x \log(x) - (1 - x) \log(1 - x)$.

source

QuantumInformation.superfidelity — Method.

superfidelity(ρ, σ)

ρ: quantum state.

σ: quantum state.

Return superfidelity between quantum states ρ and σ.

source

QuantumInformation.trace_distance — Method.

trace_distance(A, B)

A: matrix.

B: matrix.

Return trace distance between matrices A and B.

source

QuantumInformation.unres — Method.

unres(ρ)

ϕ: input matrix.

Return de-reshaping of the vector into a matrix.

source

QuantumInformation.vonneumann_entropy — Method.

vonneumann_entropy(ρ)

ρ: quantum state.

Return Von Neumann entropy of quantum state ρ.

source

QuantumInformation.werner_state — Method.

werner_state(d, α)

d: length of the vector.

α: real number from [0, 1].

Returns Werner state given by $\frac{\alpha}{d}\left(\sum_{i=0}^{\sqrt{d}-1}|ii\rangle\right) \left(\sum_{i=0}^{\sqrt{d}-1}\langle ii|\right)+ \frac{1-\alpha}{d}\sum_{i=0}^{d-1}|i\rangle\langle i|$.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: list of Kraus operators.

Transforms list of Kraus operators into dynamical matrix.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: super-operator matrix.

Transforms super-operator matrix into dynamical matrix.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: dynamical matrix.

Transforms dynamical matrix into list of Kraus operators.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: super-operator matrix.

Transforms super-operator matrix into list of Kraus operators.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: dynamical matrix.

Transforms dynamical matrix into Stinespring representation of quantum channel.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: list of Kraus operators.

Transforms list of Kraus operators into Stinespring representation of quantum channel.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: super-operator matrix.

Transforms super-operator matrix into Stinespring representation of quantum channel.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: dynamical matrix.

Transforms dynamical matrix into super-operator matrix.

source

Base.convert — Method.

convert(?, Φ)

?: type.

Φ: list of Kraus operators.

Transforms list of Kraus operators into super-operator matrix.

source